libs/naas-abi/naas_abi/apps/nexus/apps/api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nexus-api"
version = "0.0.1"
description = "NEXUS Platform API - ABI Integration Layer"
requires-python = ">=3.11"
dependencies = [
    # Core
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    # Database
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    # Redis (caching/sessions)
    "redis>=5.0.0",
    # HTTP client
    "httpx>=0.26.0",
    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.6",
    "email-validator>=2.0.0",
    # AI/LLM
    "openai>=1.10.0",
    "anthropic>=0.18.0",
    "ollama>=0.1.0",
    "tiktoken>=0.5.0",
    # Storage (S3/MinIO)
    "boto3>=1.34.0",
    # Utilities
    "structlog>=24.1.0",
    "tenacity>=8.2.0",
    "PyYAML",
    "python-socketio>=5.16.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]
cloudflare = [
    "workers-py; python_version>='3.12'",
    "workers-runtime-sdk; python_version>='3.12'",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    # Type stubs for better type checking
    "types-redis>=4.6.0.20241004",
    "types-pyyaml>=6.0.12.20250915",
    "types-boto3>=1.42.42",
    "types-python-jose>=3.5.0.20250531",
    "types-python-dateutil>=2.9.0.20260124",
]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.11"
strict = true
# Type stubs are installed for: redis, PyYAML, boto3, python-jose, python-dateutil
# FastAPI has built-in type hints (py.typed marker), no separate stub package exists
# Set to false to allow mypy to properly check FastAPI and other typed packages
ignore_missing_imports = false
# Enable pydantic plugin for better Pydantic/FastAPI type checking
# (plugin is included with pydantic package, no separate install needed)
plugins = ["pydantic.mypy"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
